Prettifies the event date/time
Adds a prettifyDate function that returns a more comfortably formatted string equivalent of a given DateTime, and uses it in EventView to make the "When?" field considerably more legible.

prettifyDate also has an optional parameter that results in the date names output in a longer format, in case such a thing is useful elsewhere.