Converted EventForm to use ListView
Replaces the Column of EventForm with a ListView (and integrates padding instead of relying on a separate object), allowing for the page to scroll. This not only fixes the renderflex overflow error when editing a text field, but it also allows for additional elements to be added in the future.