Adds time picker to event creation/addition
You can now add a time to the start and end date of any event!
KNOWN ISSUES: You can create a 1 day event with a start date after the end date, which causes an assertion error
Whenever entering the EventForm screen, the datepicker reverts to whatever is initialized for dateStart and dateEnd instead of what's in the constructor. I've highlighted this by hardcoding the dateStart/End to Dec 1.